Enable new webhook function
